+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 - BC: The `detect()` method of `CiDetector` class is no longer static.
+- BC: Rearrange namespaces, all classes are now in `CiDetector` sub-namespace.
 
 ## 1.1.0 - 2017-01-06
 - Add `getGitBranch()` method to detect Git branch of the build (supported by all CIs except TeamCity).